src/image_translator/inpainter/requirements.txt
fastapi
pillow
numpy
simple-lama-inpainting
uvicorn
python-multipart